Move StateTag and friends to runners/core-java
