+++ runners/core-java/src/main/java/org/apache/beam/runners/core/MergingStateAccessor.java
@@ -22,7 +22,6 @@ import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.util.state.State;
-import org.apache.beam.sdk.util.state.StateTag;
 
 /**
  * Interface for accessing persistent state while windows are merging.
