+++ runners/core-java/src/main/java/org/apache/beam/runners/core/NonEmptyPanes.java
@@ -24,8 +24,6 @@ import org.apache.beam.sdk.util.WindowingStrategy;
 import org.apache.beam.sdk.util.WindowingStrategy.AccumulationMode;
 import org.apache.beam.sdk.util.state.AccumulatorCombiningState;
 import org.apache.beam.sdk.util.state.ReadableState;
-import org.apache.beam.sdk.util.state.StateTag;
-import org.apache.beam.sdk.util.state.StateTags;
 
 /**
  * Tracks which windows have non-empty panes. Specifically, which windows have new elements since
