+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StateAccessor.java
@@ -20,7 +20,6 @@ package org.apache.beam.runners.core;
 import org.apache.beam.sdk.annotations.Experimental;
 import org.apache.beam.sdk.annotations.Experimental.Kind;
 import org.apache.beam.sdk.util.state.State;
-import org.apache.beam.sdk.util.state.StateTag;
 
 /**
  * Interface for accessing a {@link StateTag} in the current context.
