+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StateMerging.java
@@ -29,7 +29,6 @@ import org.apache.beam.sdk.util.state.BagState;
 import org.apache.beam.sdk.util.state.CombiningState;
 import org.apache.beam.sdk.util.state.ReadableState;
 import org.apache.beam.sdk.util.state.State;
-import org.apache.beam.sdk.util.state.StateTag;
 import org.apache.beam.sdk.util.state.WatermarkHoldState;
 import org.joda.time.Instant;
